First Iteration
This is the first iteration of the project. Currently, the user's location is tracked via the application. 
MapKit is implemented 
The next step is to highlight/draw the path of the user while they are moving.
If possible, I will attempt to draw the map in 3D, once the 2D version is complete.